Initial commit
I have finally managed to set up my speakers where I want them, but I
can clearly hear that at some positions in the room, there is a very
heavy bass, and at other positions there is almost no bass at all,
likely due to room modes. Also, some notes in bass parties are suddenly
very soft -- likely due to speaker-boundary interference -- and the bass
sounds very unclear -- likely due to reverb and long decay times at
certain frequencies.

There is a lot of pseudo-scientific information on the internet about
the physical phenomena that cause this, and I can clearly tell that it
does have a huge impact on the frequency response. But I do not have the
right equipment to measure frequency response, and moving things around
and doing things by ear is tedious.

All of these are physical phenomena, and just like with light, we can
simply simulate them. The physics are well understood, so the accuracy
of the computation only depends on how accurate the virtual
representation of the room is. I don't need something very accurate,
even a rough indication would be nice. Although one of the things that
I am curious about, is how much the level of detail matters. If I add
things like insets in corners, does that make the picture completely
different, or is the simpler version still a good approximation?

My idea is to have something with heatmap that overlays a map of the
room, that shows amplitude of the sound waves (for a given frequency).
Conversely, at a given location, we can compute the amplitude as a
function of frequency.