Include higher order reflections
This is the naive way to do it, and it is correct as far as I am aware,
but now it gets terribly terribly slow. For a single listener position
it would be doable, but if I want to do e.g. 6th order, it explodes. I
will need some kind of Monte Carlo approach.