Add random number generator
No need to depend on any external crate for this; the code is so simple
that the number of lines is probably less than the license text of an
external library.

When I get to generating floats it could become trickier, but even then,
the xoshiro authors have a very useful reference page. Also, I may be
able to get away with xoshiro128+ for floats, which should be even
faster.