Sample paths to listener, do not enumerate
I wanted to do up to 50-ish bounces, but it turns out that the factors
explode, and the range is too big for f32 to cover. This is rather
unfortunate, and it also came as a surprise to me, so I am double
checking the math, but at this point I have not found the mistake; I
think the factor should really increase by a factor num_faces at every
bounce. Then it loses a bit of energy due to the bounce (I have no idea
what a realistic value for this parameter is), and the amplitude falls
off with distance squared. But the num_faces^n factor dominates, it
grows more quickly than either 0.9^n or 1/(n^2) can shrink (assuming the
distance grows linearly with the number of bounces), so the numbers are
going to explode. Hmm...